api/requirements.txt
Flask==2.3.3
supabase==2.0.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
